Change to RecyclerView and ListAdapter

+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,13 +1,13 @@
 package ru.netology.nmedia.dto
 
 data class Post (
-    val id:Int,
-    val author:String,
-    val published:String,
-    val content:String,
-    val likes:Int=0,
-    val likedByMe:Boolean=false,
-    val shareCount:Int=0,
-    val shareByMe:Boolean=false,
-    val viewCount:Int=0
+    val id: Long,
+    val author: String,
+    val published: String,
+    val content: String,
+    val likes: Long = 0,
+    val likedByMe: Boolean = false,
+    val shareCount: Long = 0,
+    val shareByMe: Boolean = false,
+    val viewCount: Long = 0
 )
